Change to random quiz link/redirect
Instead of having to fetch a random quiz to show on the landing page we link to /randomquiz.
When /randomquiz gets hit we query the DB for every quiz and redirect to a random one. This is being done in getServerSideProps.

Removed /api/randomquiz since we only need the randomquiz pages route now.